board:2093 bugprone-signed-bitwise reaches ZERO, and the control caught a truncation I wrote
  this check       112 -> 0
  findings        6762 -> 6658
  Venice      c713d304   unchanged
  OldTown     b0f9f7d7   unchanged

THREE KINDS OF FINDING WERE IN THE 112, AND ONLY ONE WAS OURS.

Most were SDL's own macros. SDL_GPU_BUFFERUSAGE_VERTEX is (1u << 0) and the shift amount is a signed
literal, so every OR of two usage flags was reported at our line for something written in SDL's
header. That is not the hazard the check exists for -- a signed VARIABLE in a bitwise expression,
where the sign bit is a trap -- and the check ships an option that says so:
IgnorePositiveIntegerLiterals. It is set, with the count and the reason on the line above it, and
the half that matters stays on.

The rest were ours and each needed a decision rather than a cast. A Unicode code point is parsed as
a long so the range check can catch a negative, and is then CARRIED as uint32_t, which is what it
is. Sha256's rotate takes an unsigned shift amount. A bit width like kCoarseDrop is declared
unsigned rather than cast at each use. A half-float is widened once instead of three times. And
Network::KeyAt shifted an int64_t left by 32, which is undefined for a large row; it now composes
the key in uint64_t and converts once at the end.

AND THE CONTROL EARNED ITS KEEP. A script wrapped the flagged operand in static_cast<uint32_t>,
which is right for a shift AMOUNT and silently wrong for a VALUE: `long cx = hx >> kCoarseDrop`
became `static_cast<uint32_t>(hx) >> kCoarseDrop` and threw away the top 32 bits of a tile
coordinate. The build was clean. The tests were clean. Venice moved, and that is the only reason it
was found. Repaired to the same width, and Venice returned to c713d304 on the next render.

include/Scene.h is fixed rather than skipped, at the owner's word that the public interface must be
correct too: four uint8_t role bits OR-ed together promote to int, so each is widened explicitly
before the OR and the result is narrowed once, which is what was always meant.

+++ src/base/io/Png.cpp
@@ -141,7 +141,7 @@ Png ReadPng(const uint8_t *bytes, size_t length) {
           return Refuse("row " + std::to_string(row) + " declares filter " +
                         std::to_string(static_cast<int>(filter)) + ", and PNG has five");
       }
-      outRow[byte] = static_cast<uint8_t>(value & 0xff);
+      outRow[byte] = static_cast<uint8_t>(static_cast<uint32_t>(value) & 0xff);
     }
   }
